Refactor usage of executors
